+++ modules/cpu/render/MaterialShared.h
@@ -78,9 +78,8 @@ struct Material
   Material_GetBSDFFunc getBSDF;
   Material_GetTransparencyFunc getTransparency;
   Material_SelectNextMediumFunc selectNextMedium;
-  vec3f emission; // simple constant (spatially and angular) emission, returns
-                  // radiance; TODO SV-EDFs
-  TextureParam emissionMap;
+  vec3f emission; // angular constant, returns radiance
+  TextureParam emissionMap; // spatially varying emission (SV-EDF)
 
   MicrofacetAlbedoTables *microfacetAlbedoTables;