+++ modules/cpu/render/materials/Principled.cpp
@@ -74,9 +74,9 @@ void Principled::commit()
   float outsideTransmissionDepth =
       getParam<float>("outsideTransmissionDepth", 1.f);
 
-  // emissiveColor can be constant and/or texture to modulate the radiance,
-  // intensity is a constant factor
-  MaterialParam3f emissiveColor = getMaterialParam3f("emissiveColor", 0.f);
+  MaterialParam3f emissiveColor =
+      getMaterialParam3f("emissiveColor", vec3f(0.f));
+  // intensity is an additional constant factor to emission
   emissiveColor.factor *= getParam<float>("intensity", 1.f);
   getSh()->super.emission = emissiveColor.factor;
   getSh()->super.emissionMap = emissiveColor.tex;